requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
google-generativeai==0.8.3
openai>=1.0.0

# RAG pipeline
chromadb>=0.5.0
PyPDF2>=3.0.0
python-docx>=1.1.0
